inference/requirements.txt
fastapi[standard]
python-logstash
